requirements.txt
asgiref==3.2.10
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
boto3==1.14.20
botocore==1.17.20
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0.8
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-developer-panel==0.1.2
django-filter==1.1.0
django-htmlmin==0.11.0
django-import-export==2.5.0
django-on-heroku==1.1.2
django-pdb==0.6.2
django-pyodbc==1.1.3
django-template-debug==0.3.5
django-utils-six==2.0
django-versatileimagefield==2.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
docutils==0.15.2
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
importlib-metadata==4.6.1
jmespath==0.10.0
libmagic==1.0
MarkupPy==1.14
mysql-connector-python==8.0.26
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
platformdirs==2.0.2
protobuf==3.17.3
psycopg2-binary==2.9.1
pyodbc==4.0.31
python-dateutil==2.8.1
python-decouple==3.4
python-magic-bin==0.4.14
python-magic-win64==0.4.13
pytz==2020.1
PyYAML==5.4.1
s3transfer==0.3.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
tablib==3.0.0
typing-extensions==3.10.0.0
urllib3==1.25.9
virtualenv==20.6.0
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
